New tests

+++ client/src/commonMain/kotlin/com/lightningkite/lightningserver/db/ClientModelRestEndpointsMock.kt
@@ -1,5 +1,6 @@
 package com.lightningkite.lightningserver.db
 
+import com.lightningkite.kiteui.ConnectionException
 import com.lightningkite.kiteui.Console
 import com.lightningkite.kiteui.ConsoleRoot
 import com.lightningkite.lightningdb.AggregateQuery
@@ -28,6 +29,8 @@ open class ClientModelRestEndpointsMock<T : HasId<ID>, ID : Comparable<ID>>(
     val delayAmount: Duration = 0.1.seconds,
     val log: Console? = null,
 ) : ClientModelRestEndpoints<T, ID> {
+    open var connectivityFailure: Boolean = false
+
     val data = HashMap<ID, T>()
     open fun change(collectionUpdates: CollectionUpdates<T, ID>) {
         log?.log("changes: $collectionUpdates")
@@ -41,6 +44,7 @@ open class ClientModelRestEndpointsMock<T : HasId<ID>, ID : Comparable<ID>>(
 
     override suspend fun default(): T = throw IllegalArgumentException()
     override suspend fun query(input: Query<T>): List<T> {
+        if(connectivityFailure) throw ConnectionException("Dead")
         log?.log("query")
         delay(delayAmount)
         return data.values
@@ -55,38 +59,45 @@ open class ClientModelRestEndpointsMock<T : HasId<ID>, ID : Comparable<ID>>(
 
     override suspend fun queryPartial(input: QueryPartial<T>): List<Partial<T>> = TODO()
     override suspend fun detail(id: ID): T {
+        if(connectivityFailure) throw ConnectionException("Dead")
         log?.log("detail")
         delay(delayAmount)
         return data[id] ?: throw LsErrorException(404, LSError(404, "not-found", "", ""))
     }
     override suspend fun insertBulk(input: List<T>): List<T> {
+        if(connectivityFailure) throw ConnectionException("Dead")
         log?.log("insertBulk")
         delay(delayAmount)
         return input.onEach { change(CollectionUpdates(updates = setOf(it))) }
     }
 
     override suspend fun insert(input: T): T {
+        if(connectivityFailure) throw ConnectionException("Dead")
         log?.log("insert")
         delay(delayAmount)
         return input.also { change(CollectionUpdates(updates = setOf(it))) }
     }
     override suspend fun upsert(id: ID, input: T): T {
+        if(connectivityFailure) throw ConnectionException("Dead")
         log?.log("upsert")
         delay(delayAmount)
         return input.also { change(CollectionUpdates(updates = setOf(it))) }
     }
     override suspend fun bulkReplace(input: List<T>): List<T> {
+        if(connectivityFailure) throw ConnectionException("Dead")
         log?.log("bulkReplace")
         delay(delayAmount)
         return input.onEach { change(CollectionUpdates(updates = setOf(it))) }
     }
 
     override suspend fun replace(id: ID, input: T): T {
+        if(connectivityFailure) throw ConnectionException("Dead")
         log?.log("replace")
         delay(delayAmount)
         return input.also { change(CollectionUpdates(updates = setOf(it))) }
     }
     override suspend fun bulkModify(input: MassModification<T>): Int {
+        if(connectivityFailure) throw ConnectionException("Dead")
         log?.log("bulkModify")
         delay(delayAmount)
         return data.values.toList().count {
@@ -98,6 +109,7 @@ open class ClientModelRestEndpointsMock<T : HasId<ID>, ID : Comparable<ID>>(
     }
 
     override suspend fun modifyWithDiff(id: ID, input: Modification<T>): EntryChange<T> {
+        if(connectivityFailure) throw ConnectionException("Dead")
         log?.log("modifyWithDiff")
         delay(delayAmount)
         return EntryChange(
@@ -106,6 +118,7 @@ open class ClientModelRestEndpointsMock<T : HasId<ID>, ID : Comparable<ID>>(
     }
 
     override suspend fun modify(id: ID, input: Modification<T>): T {
+        if(connectivityFailure) throw ConnectionException("Dead")
         log?.log("modify")
         delay(delayAmount)
         return data[id]?.let { input(it) }?.also { change(CollectionUpdates(updates = setOf(it))) }
@@ -116,6 +129,7 @@ open class ClientModelRestEndpointsMock<T : HasId<ID>, ID : Comparable<ID>>(
     }
 
     override suspend fun bulkDelete(input: Condition<T>): Int {
+        if(connectivityFailure) throw ConnectionException("Dead")
         log?.log("bulkDelete")
         delay(delayAmount)
         return data.values.toList().count {
@@ -127,12 +141,14 @@ open class ClientModelRestEndpointsMock<T : HasId<ID>, ID : Comparable<ID>>(
     }
 
     override suspend fun delete(id: ID): Unit {
+        if(connectivityFailure) throw ConnectionException("Dead")
         log?.log("delete")
         delay(delayAmount)
         change(CollectionUpdates(remove = setOf(id)))
     }
 
     override suspend fun count(input: Condition<T>): Int {
+        if(connectivityFailure) throw ConnectionException("Dead")
         log?.log("count")
         delay(delayAmount)
         return data.values.asSequence().filter { input(it) }.count { input(it) }
